requirements.txt
altair==5.5.0
annotated-types==0.7.0
attrs==24.2.0
blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
chardet==3.0.4
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
filelock==3.16.1
fpdf==1.7.2
fsspec==2024.10.0
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.10
google-api-core==2.23.0
google-api-python-client==2.154.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.66.0
googletrans==4.0.0rc1
grpcio==1.68.0
grpcio-status==1.68.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.12.1
httpcore==0.9.1
httplib2==0.22.0
httpx==0.13.3
huggingface-hub==0.26.3
hyperframe==5.2.0
idna==2.10
intel-openmp==2021.4.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langid==1.1.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mkl==2021.4.0
mpmath==1.3.0
narwhals==1.15.0
networkx==3.4.2
numpy==2.1.3
packaging==24.2
pandas==2.2.3
pdf2image==1.17.0
pillow==11.0.0
proto-plus==1.25.0
protobuf==5.29.0
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.2
pydantic_core==2.27.1
pydeck==0.9.1
Pygments==2.18.0
pyparsing==3.2.0
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
regex==2024.11.6
reportlab==4.2.5
requests==2.32.3
rfc3986==1.5.0
rich==13.9.4
rpds-py==0.21.0
rsa==4.9
safetensors==0.4.5
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
streamlit==1.40.2
sympy==1.13.1
tbb==2021.13.1
tenacity==9.0.0
tokenizers==0.20.3
toml==0.10.2
torch==2.3.1
tornado==6.4.2
tqdm==4.67.1
transformers==4.46.3
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
watchdog==6.0.0
